+++ supertokens_python/recipe/emailpassword/interfaces.py
@@ -25,6 +25,7 @@ if TYPE_CHECKING:
     from supertokens_python.framework import BaseRequest, BaseResponse
     from .utils import EmailPasswordConfig
     from .types import User, UsersResponse, FormField
+    from supertokens_python.recipe.session import Session
 
 
 class SignUpResult(ABC):
@@ -98,7 +99,7 @@ class CreateResetPasswordWrongUserIdErrorResult(CreateResetPasswordResult):
 
 class ResetPasswordUsingTokenResult(ABC):
     def __init__(self, status: Literal['OK',
-                 'RESET_PASSWORD_INVALID_TOKEN_ERROR']):
+                                       'RESET_PASSWORD_INVALID_TOKEN_ERROR']):
         self.status = status
         self.is_ok = False
         self.is_reset_password_invalid_token_error = False
@@ -111,8 +112,7 @@ class ResetPasswordUsingTokenOkResult(ResetPasswordUsingTokenResult):
         self.is_reset_password_invalid_token_error = False
 
 
-class ResetPasswordUsingTokenWrongUserIdErrorResult(
-        ResetPasswordUsingTokenResult):
+class ResetPasswordUsingTokenWrongUserIdErrorResult(ResetPasswordUsingTokenResult):
     def __init__(self):
         super().__init__('RESET_PASSWORD_INVALID_TOKEN_ERROR')
         self.is_ok = False
@@ -136,8 +136,7 @@ class UpdateEmailOrPasswordOkResult(UpdateEmailOrPasswordResult):
         self.is_unknown_user_id_error = False
 
 
-class UpdateEmailOrPasswordEmailAlreadyExistsErrorResult(
-        UpdateEmailOrPasswordResult):
+class UpdateEmailOrPasswordEmailAlreadyExistsErrorResult(UpdateEmailOrPasswordResult):
     def __init__(self):
         super().__init__('EMAIL_ALREADY_EXISTS_ERROR')
         self.is_ok = False
@@ -145,8 +144,7 @@ class UpdateEmailOrPasswordEmailAlreadyExistsErrorResult(
         self.is_unknown_user_id_error = False
 
 
-class UpdateEmailOrPasswordUnknownUserIdErrorResult(
-        UpdateEmailOrPasswordResult):
+class UpdateEmailOrPasswordUnknownUserIdErrorResult(UpdateEmailOrPasswordResult):
     def __init__(self):
         super().__init__('UNKNOWN_USER_ID_ERROR')
         self.is_ok = False
@@ -171,7 +169,8 @@ class RecipeInterface(ABC):
         pass
 
     @abstractmethod
-    async def reset_password_using_token(self, token: str, new_password: str, user_context: any) -> ResetPasswordUsingTokenResult:
+    async def reset_password_using_token(self, token: str, new_password: str,
+                                         user_context: any) -> ResetPasswordUsingTokenResult:
         pass
 
     @abstractmethod
@@ -283,16 +282,14 @@ class GenerateEmailVerifyTokenPostResponse(ABC):
         }
 
 
-class GenerateEmailVerifyTokenPostOkResponse(
-        GenerateEmailVerifyTokenPostResponse):
+class GenerateEmailVerifyTokenPostOkResponse(GenerateEmailVerifyTokenPostResponse):
     def __init__(self):
         super().__init__('OK')
         self.is_ok = True
         self.is_email_already_verified_error = False
 
 
-class GenerateEmailVerifyTokenPostEmailAlreadyVerifiedErrorResponse(
-        GenerateEmailVerifyTokenPostResponse):
+class GenerateEmailVerifyTokenPostEmailAlreadyVerifiedErrorResponse(GenerateEmailVerifyTokenPostResponse):
     def __init__(self):
         super().__init__('EMAIL_ALREADY_VERIFIED_ERROR')
         self.is_ok = False
@@ -326,15 +323,14 @@ class GeneratePasswordResetTokenPostResponse(ABC):
         }
 
 
-class GeneratePasswordResetTokenPostOkResponse(
-        GeneratePasswordResetTokenPostResponse):
+class GeneratePasswordResetTokenPostOkResponse(GeneratePasswordResetTokenPostResponse):
     def __init__(self):
         super().__init__('OK')
 
 
 class PasswordResetPostResponse(ABC):
     def __init__(self, status: Literal['OK',
-                 'RESET_PASSWORD_INVALID_TOKEN_ERROR']):
+                                       'RESET_PASSWORD_INVALID_TOKEN_ERROR']):
         self.status = status
 
     def to_json(self):
@@ -355,12 +351,15 @@ class PasswordResetPostInvalidTokenResponse(PasswordResetPostResponse):
 
 class SignInPostResponse(ABC):
     def __init__(
-            self, status: Literal['OK', 'WRONG_CREDENTIALS_ERROR'], user: Union[User, None]):
+            self, status: Literal['OK', 'WRONG_CREDENTIALS_ERROR'],
+            user: Union[User, None] = None,
+            session: Union[Session, None] = None):
         self.type = 'emailpassword'
         self.is_ok = False
         self.is_wrong_credentials_error = False
         self.status = status
         self.user = user
+        self.session = session
 
     def to_json(self):
         response = {
@@ -379,25 +378,28 @@ class SignInPostResponse(ABC):
 
 
 class SignInPostOkResponse(SignInPostResponse):
-    def __init__(self, user: User):
-        super().__init__('OK', user)
+    def __init__(self, user: User, session: Session):
+        super().__init__('OK', user, session)
         self.is_ok = True
 
 
 class SignInPostWrongCredentialsErrorResponse(SignInPostResponse):
     def __init__(self):
-        super().__init__('WRONG_CREDENTIALS_ERROR', None)
+        super().__init__('WRONG_CREDENTIALS_ERROR')
         self.is_wrong_credentials_error = True
 
 
 class SignUpPostResponse(ABC):
     def __init__(
-            self, status: Literal['OK', 'EMAIL_ALREADY_EXISTS_ERROR'], user: Union[User, None]):
+            self, status: Literal['OK', 'EMAIL_ALREADY_EXISTS_ERROR'],
+            user: Union[User, None] = None,
+            session: Union[Session, None] = None):
         self.type = 'emailpassword'
         self.is_ok = False
         self.is_email_already_exists_error = False
         self.status = status
         self.user = user
+        self.session = session
 
     def to_json(self):
         response = {
@@ -416,14 +418,14 @@ class SignUpPostResponse(ABC):
 
 
 class SignUpPostOkResponse(SignUpPostResponse):
-    def __init__(self, user: User):
-        super().__init__('OK', user)
+    def __init__(self, user: User, session: Session):
+        super().__init__('OK', user, session)
         self.is_ok = True
 
 
 class SignUpPostEmailAlreadyExistsErrorResponse(SignUpPostResponse):
     def __init__(self):
-        super().__init__('EMAIL_ALREADY_EXISTS_ERROR', None)
+        super().__init__('EMAIL_ALREADY_EXISTS_ERROR')
         self.is_email_already_exists_error = True